listendc/requirements.txt
discord.py>=2.0.0
discord.py-self>=1.9.0
aiohttp>=3.8.0
paho-mqtt>=1.6.0
pyyaml>=6.0
